TestLevelFlag: Don't print to stderr
If you run `go test` in the root of Zap right now,
the output includes:

    invalid value "unknown" for flag -level: unrecognized level: "unknown"
    Usage of test:
      -level value
            minimum enabled logging level

This comes from TestLevelFlag which instantiates a flag.FlagSet.
By default, flag.FlagSet will print messages to stderr,
which is why we see this output.

Switch it to discard its output like other FlagSets
instantiated in other similar tests in the same file.