WrapError: wrap an error with fields to be logged by zap.Error
Related to https://github.com/uber-go/guide/pull/179

Callsites that receive an error should either log, or return an error.

However, if the callsite has additioanl context, the simplest option is
to add it to the error, but it's then flattened into a string, losing
the benefit of structured logging. This often results in callsites
logging with additional fields, and returning an error that is likely
to be logged again.

`WrapError` provides a way for callsites to return an error that
includes fields to be logged, which will be added to an `errorFields`
key.