Clarify argument capturing behaviour of With
Make it clear that the contract with the With and logging methods is such that
any arguments that could change are captured at the time of the With and logging
method, and that any other behaviour introduced in future is intentionally
contrary to this contract.

There are proposed (https://github.com/uber-go/zap/pull/1319) future opt-in,
changes that will explicitly relax this contract.