Add benchmarks for chained Withs
Add benchmarks to show the performance of chaining child loggers via 'With' and
whether the logger is used or not used.

This is intended to establish a baseline for a proposed optimisation in PR
https://github.com/uber-go/zap/pull/1319, in particular, we would like to see
if we can avoid the expense of With() if a logger is never used.

Currently, the performance is similar irrespective of whether the logger is used
or not:

```
 % go test -bench=Benchmark5Withs
goos: darwin
goarch: arm64
pkg: go.uber.org/zap
Benchmark5WithsUsed-10       	  512342	      2276 ns/op
Benchmark5WithsNotUsed-10    	  570768	      2201 ns/op
```