chore(CHANGELOG): Move reference links to sections
Some of the links in the changelog entries
are at the bottom of the changelog file
while others are in the section where they're used.

Adding these to the bottom of the file is unmanageable.
They should be added to the section where they're used.

This moves all remaining links from the bottom of the file
to the sections where they're used.