lint: Pass zapcore and zaptest packages
golangci-lint running revive currently fails on master. Exempt the dot-import
failures since these are used to reduce a large number of qualifiers in the
zapcore packages's _test package tests.

This change also simplifies an if/else statement with a logical equivalent and
standardizes a test that imported a package and a dot-import of the same
package simultaneously.