Rename license file to satisfy go docs for nested modules (#1389)
The documentation for module `uber.go/zap/exp` is not eligible to be
rendered on pkg.go.dev as no license is detected. For example,
https://pkg.go.dev/go.uber.org/zap/exp@v0.2.0/zapslog will not render.

This is due to some possibly unfortunate behavior in the Go command that
special cases copying a `LICENSE` file from the repository for root for
any nested modules lacking a `LICENSE` file. This is problematic for Zap
as Zap's license is in a file named `LICENSE.txt`. Renaming the license
file to `LICENSE` remedies the issue.

Relevant issues:
- https://github.com/golang/go/issues/43817#issuecomment-764215261
- https://github.com/golang/go/issues/44758#issuecomment-789908134