fix: WithGroup.WithGroup.Info should not create an empty group
This requires holding onto a slice instead of a single string
to create the nested namespaces.

The slice is cloned on each WithGroup call.
We could probably do this more efficiently with an immutable tree of
references, but deferring that for now.

In the process, also make the addedNamespace business
append during the record.Attrs loop instead of after.
This is more efficient because we don't have to shift the slice over
afterwards.