remove redundant capitalizations in unmarshalText
unmarshalText currently checks for almost all of the levels as
the configured string values as well as their capitalized versions.

But we call unmarshalText with bytes.ToLower() when we actually parse
the levels, so we do not need to check for the capitalized versions
separately since this is redundant.

Previous version of this PR was trying to add "WARNING" as a check
to this since that was recently added but the corresponding capitalized
string wasn't checked, but @abhinav pointed out that we actually call
this with ToLower, so it's better to remove the redundant checks.