Safer behaviour of (mis)logging badly coded marshallers
ArrayMarshalerType, ObjectMarshalerType, InlineMarshalerType all call
user provided methods on user provided values.
They may fail/panic and have to be guarded.

Use guard similar to the one for stringer with the following caveats:

- ObjectMarshalerType in case of nil-ness will be logged as string rather than map.
- ArrayMarshalerType - we have to insert a synthetic empty marshaller to log an empty slice.
  Otherwise the default shape of map shines through
- InlineMarshalerType - in case of any error it will be de-inlined to show exact cause of error.
  Otherwise error will be attached to the wrapping object which will make debugging and fixing harder

While we introduce enough coverage, the test harness is not using currently subtests.
This patch does not address this to keep changes to a minimum and give time to discuss if this is necessary.
Thus there are 3 tests that were changed - the error cases, the success cases and inline case (separate).