Update lazy logger not to materialize unless it's being written to
Lazy logger is frequently used to delay (expensive) logger initialization unless we actually write to it.

This works for a pattern like:
```lang=go
logger := logger.With(.../*fields)

err := foo()
if err != nil {
   logger.Info("foo: %w, err)
}
```

Another common pattern though might be
```lang=go
logger := logger.With(.../*fields)

// a lot of code

logger.Debug("something happened")
```

Here, despite the logs being discarded, we still initialize the logger even though it might be expensive.

This diff attempts to mitigate this by using the original core for level checking. This, afaiu, should be safe as `With` is only for field creation, and should not be changing log levels.

Preserving the original core to avoid race conditions (torn interfaces) on actual `With` initialization.